Thread testem-electron.js config file through to TestTask
